refactor: make encode_path private
