+++ lib/carrierwave/uploader/url.rb
@@ -5,23 +5,7 @@ module CarrierWave
     module Url
       extend ActiveSupport::Concern
       include CarrierWave::Uploader::Configuration
-
-      module ClassMethods
-        def encode_path(path)
-          # based on Ruby < 2.0's URI.encode
-          safe_string = URI::REGEXP::PATTERN::UNRESERVED + '\/'
-          unsafe = Regexp.new("[^#{safe_string}]", false)
-
-          path.gsub(unsafe) do
-            us = $&
-            tmp = ''
-            us.each_byte do |uc|
-              tmp << sprintf('%%%02X', uc)
-            end
-            tmp
-          end
-        end
-      end
+      include CarrierWave::Utilities::Uri
 
       ##
       # === Parameters
@@ -36,7 +20,7 @@ module CarrierWave
         if file.respond_to?(:url) and not file.url.blank?
           file.method(:url).arity == 0 ? file.url : file.url(options)
         elsif file.respond_to?(:path)
-          path = self.class.encode_path(file.path.gsub(File.expand_path(root), ''))
+          path = encode_path(file.path.gsub(File.expand_path(root), ''))
 
           if host = asset_host
             if host.respond_to? :call
