+++ lib/carrierwave/utilities.rb
@@ -1,0 +1,8 @@
+# encoding: utf-8
+
+require 'carrierwave/utilities/uri'
+
+module CarrierWave
+  module Utilities
+  end
+end
